+++ lib/src/widget.dart
@@ -31,12 +31,18 @@ abstract class SDUIWidget {
     return this;
   }
 
-  Widget? toIcon(String? code, {double? size = 16.0, String? color}) {
+  Widget? toIcon(String? code, {double? size, String? color}) {
     if (code?.startsWith("http://") == true ||
         code?.startsWith("https://") == true) {
-      return Image.network(code!, width: size, height: size,);
+      return Image.network(
+        code!,
+        width: size,
+        height: size,
+      );
     } else {
-      return code == null ? null : Icon(toIconData(code), size: size, color: toColor(color));
+      return code == null
+          ? null
+          : Icon(toIconData(code), size: size, color: toColor(color));
     }
   }